+++ src/main/java/com/gdsc/knu/repository/FriendRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FriendRepository extends JpaRepository<Friend, Integer> {
-    List<Friend> findByUserId(Integer userId);
-    List<Friend> findByFriendUserId(Integer friendUserId);
+public interface FriendRepository extends JpaRepository<Friend, Long> {
+    List<Friend> findByUserId(Long userId);
+    List<Friend> findByFriendUserId(Long friendUserId);
 
-    Optional<Friend> findByUserIdAndFriendUserId(Integer userId, Integer friendUserId);
+    Optional<Friend> findByUserIdAndFriendUserId(Long userId, Long friendUserId);
 }
